Implement method to retrieve the actual job instead of the name
Closes #1166

+++ src/main/java/com/gamingmesh/jobs/commands/list/leaveall.java
@@ -36,7 +36,7 @@ public class leaveall implements Cmd {
 		Util.LEAVECONFIRM.add(uuid);
 
 		plugin.getServer().getScheduler().runTaskLater(plugin, () -> Util.LEAVECONFIRM.remove(uuid),
-		    20 * Jobs.getGCManager().ConfirmExpiryTime);
+		    (long) (20 * Jobs.getGCManager().ConfirmExpiryTime));
 
 		pSender.sendMessage(Jobs.getLanguage().getMessage("command.leaveall.confirmationNeed", "[time]",
 		    Jobs.getGCManager().ConfirmExpiryTime));
